refactor: Home 컴포넌트를 screens에서 components로 이동
- src/screens/home/components/에서 src/components/home/로 컴포넌트 이동
- home_screen.tsx와 home_style.ts 업데이트하여 새로운 경로 반영
