+++ src/components/home/greeting_section/index.tsx
