+++ src/screens/home/home_screen.tsx
@@ -1,78 +1,22 @@
-import { useFocusEffect } from '@react-navigation/native';
 import { router } from 'expo-router';
 import * as SplashScreen from 'expo-splash-screen';
-import { useCallback, useEffect, useRef } from 'react';
-import { ScrollView, View, ActivityIndicator, Alert } from 'react-native';
+import { useCallback, useEffect } from 'react';
+import { ScrollView, View, ActivityIndicator } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
+import BenefitsSection from '@/src/components/home/benefit_section';
+import HomeHeader from '@/src/components/home/home_header';
+import RecommendedMatchCard from '@/src/components/home/recommended_match_card';
+import TodayMatchStatus from '@/src/components/home/today_match_status';
 import { ROUTES } from '@/src/constants/routes';
-import { useUserProfile, useLogout } from '@/src/hooks/queries';
-import BenefitsSection from '@/src/screens/home/components/benefit_section';
-import HomeHeader from '@/src/screens/home/components/home_header';
-import RecommendedMatchCard from '@/src/screens/home/components/recommended_match_card';
-import TodayMatchStatus from '@/src/screens/home/components/today_match_status';
+import { useUserProfile, useTeam } from '@/src/hooks/queries';
 import { styles } from '@/src/screens/home/home_style';
 import { theme } from '@/src/theme';
 
 export default function HomeScreen() {
   const insets = useSafeAreaInsets();
-  const { data: userProfile, isLoading, refetch, error } = useUserProfile();
-  const logoutMutation = useLogout();
-  const alertShownRef = useRef(false);
-
-  const handleLogoutAndRedirect = useCallback(async () => {
-    try {
-      await logoutMutation.mutateAsync();
-    } catch (error) {
-      console.error('로그아웃 처리 중 오류:', error);
-    }
-  }, [logoutMutation]);
-
-  const handleErrorAlert = useCallback(() => {
-    if (alertShownRef.current) {
-      return;
-    }
-
-    alertShownRef.current = true;
-
-    Alert.alert(
-      '인증 오류',
-      '로그인이 필요합니다. 다시 로그인해주세요.',
-      [
-        {
-          text: '로그인 하기',
-          onPress: () => {
-            alertShownRef.current = false;
-            handleLogoutAndRedirect();
-          },
-        },
-      ],
-      {
-        cancelable: false,
-        onDismiss: () => {
-          alertShownRef.current = false;
-        },
-      }
-    );
-  }, [handleLogoutAndRedirect]);
-
-  useFocusEffect(
-    useCallback(() => {
-      refetch();
-    }, [refetch])
-  );
-
-  useEffect(() => {
-    if (error) {
-      handleErrorAlert();
-    }
-  }, [error, handleErrorAlert]);
-
-  useEffect(() => {
-    if (!isLoading && !error) {
-      SplashScreen.hideAsync();
-    }
-  }, [isLoading, error]);
+  const { data: userProfile, isLoading } = useUserProfile();
+  const { data: team } = useTeam(userProfile?.teamId || 0);
 
   const handleMatchPress = useCallback((matchDate?: string) => {
     if (matchDate) {
@@ -110,10 +54,16 @@ export default function HomeScreen() {
           { paddingBottom: insets.bottom + theme.spacing.spacing1 },
         ]}
       >
-        <TodayMatchStatus teamId={userProfile?.teamId || null} />
+        <TodayMatchStatus
+          teamId={userProfile?.teamId || null}
+          team={team || null}
+        />
 
         <View style={styles.matchSection}>
-          <RecommendedMatchCard onMatchPress={handleMatchPress} />
+          <RecommendedMatchCard
+            onMatchPress={handleMatchPress}
+            teamId={userProfile?.teamId || null}
+          />
         </View>
 
         <View style={styles.serviceSection}>
